node-agent/model-runner/requirements.txt
llama-cpp-python
fastapi
uvicorn
pydantic
requests
tqdm
python-dotenv
aiofiles
